Implement update, create, and delete endpoints for questions.
