+++ Application/ApplicationAssembly.cs
@@ -1,0 +1,3 @@
+namespace Application;
+
+public class ApplicationAssembly { }
