+++ Application/Contracts/IUtilityService.cs
@@ -1,0 +1,6 @@
+namespace Application.Contracts;
+
+public interface IUtilityService
+{
+    string NormalizeText(string value);
+}
