+++ Application/Services/AuthenticationService.cs
@@ -1,6 +1,7 @@
 using System.IdentityModel.Tokens.Jwt;
 using System.Security.Claims;
 using System.Text;
+using Application.Contracts;
 using Domain.Constants;
 using Domain.DTO.Request;
 using Domain.DTO.Response;
