--- Application/Services/IUtilityService.cs
@@ -1,7 +1,0 @@
-namespace Application.Services
-{
-    public interface IUtilityService
-    {
-        string NormalizeText(string value);
-    }
-}
