+++ Application/Services/ServiceManager.cs
@@ -1,3 +1,4 @@
+using Application.Contracts;
 using AutoMapper;
 
 namespace Application.Services;
