+++ Application/Services/TagService.cs
@@ -1,3 +1,4 @@
+using Application.Contracts;
 using Domain.Entities;
 using Infrastructure.Repositories;
