+++ Application/Services/UtilityService.cs
@@ -1,5 +1,6 @@
 using System.Text;
 using System.Text.RegularExpressions;
+using Application.Contracts;
 
 namespace Application.Services;
