--- Presentation/AssemblyReference.cs
@@ -1,2 +1,0 @@
-namespace Presentation;
-public static class AssemblyReference { }
