+++ Presentation/PresentationAssembly.cs
@@ -1,0 +1,3 @@
+namespace Presentation;
+
+public static class PresentationAssembly { }
